services/analytics-service/requirements.txt
# Analytics Service Dependencies

# Core Framework
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.8.0
pydantic-settings==2.4.0

# Database
sqlalchemy==2.0.35
alembic==1.13.0
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Redis
redis[hiredis]==5.0.1

# Message Queue
aio-pika==9.4.0

# Authentication
python-jose[cryptography]==3.3.0

# HTTP Client
httpx==0.27.0

# OpenAI (RAG Ask Intelligence feature)
openai>=1.0.0

# Logging
structlog==24.4.0

# Utilities
python-dateutil==2.8.2
pytz==2023.4

# Container Management
docker>=7.0.0

# Testing
pytest==8.2.0
pytest-asyncio==0.24.0
pytest-cov==4.1.0
pytest-mock==3.12.0

# Testing

# Development

# Service-specific dependencies
# Data analysis
pandas==2.1.3
numpy==1.26.2

# Intelligence Features
pybursts>=0.1.1

# Visualization
plotly==5.18.0
matplotlib==3.8.2

# Templating
jinja2==3.1.2

# PDF generation
weasyprint==60.1

# Celery
celery==5.3.4

# Rate Limiting
slowapi==0.1.9

# WebSocket
websockets==12.0
